refactor: 사용하지 않는 파일 제거, 테스트케이스를 위해서 DataBase에 clear 메서드 추가
