+++ src/main/java/db/DataBase.java
@@ -8,7 +8,9 @@ import com.google.common.collect.Maps;
 import model.User;
 
 public class DataBase {
-    private static Map<String, User> users = Maps.newHashMap();
+    private DataBase() {}
+
+    private static final Map<String, User> users = Maps.newHashMap();
 
     public static void addUser(User user) {
         users.put(user.getUserId(), user);
@@ -24,5 +26,9 @@ public class DataBase {
 
     public static Collection<User> findAll() {
         return users.values();
+    }
+
+    public static void clear() {
+        users.clear();
     }
 }
