+++ src/main/java/webserver/http/request/handler/UserListRequestHandler.java
@@ -1,5 +1,9 @@
 package webserver.http.request.handler;
 
+import java.io.IOException;
+
+import model.UserList;
+
 public class UserListRequestHandler implements RequestHandler {
     static final String REQUEST_INDEX = "/user/list";
 
@@ -7,7 +11,8 @@ public class UserListRequestHandler implements RequestHandler {
         return REQUEST_INDEX;
     }
     @Override
-    public byte[] execute() {
-        return "".getBytes();
+    public byte[] execute() throws IOException {
+        UserList userList = new UserList();
+        return userList.toUserList().getBytes();
     }
 }
