+++ src/main/java/webserver/http/request/handler/exception/InvalidRequestException.java
@@ -1,4 +1,4 @@
-package webserver.http.service.exception;
+package webserver.http.request.handler.exception;
 
 public class InvalidRequestException extends RuntimeException{
     public InvalidRequestException(String message) {
